pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "thgkt"
version = "0.1.0"
description = "Temporal Heterogeneous Graph Neural Networks for Knowledge Tracing and Concept Dependency Modeling"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "OpenAI Codex" }
]
dependencies = [
  "torch>=2.9,<3",
  "torch-geometric>=2.7",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]
ml = [
  "pandas>=2.2",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-p no:cacheprovider"
